dev-requirements.txt
autopep8==1.5.2
click==7.1.2
hiredis==1.0.1
mysqlclient==1.4.6
pkg-resources==0.0.0
pycodestyle==2.6.0
python-decouple==3.3
pytz==2020.1
redis==3.5.1
rq==1.3.0
SQLAlchemy==1.3.16
